requirements.txt
fastapi==0.129.0
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
starlette==0.52.1